+++ ui/src/components/credentials/LinkTable.tsx
@@ -24,7 +24,7 @@ import { ReactComponent as IconLink } from "src/assets/icons/link-03.svg";
 import { ErrorResult } from "src/components/shared/ErrorResult";
 import { NoResults } from "src/components/shared/NoResults";
 import { TableCard } from "src/components/shared/TableCard";
-import { useEnvContext } from "src/contexts/env.context";
+import { useEnvContext } from "src/contexts/env";
 import { APIError, processZodError } from "src/utils/adapters";
 import { isAbortedError, makeRequestAbortable } from "src/utils/browser";
 import { ACCESSIBLE_UNTIL, LINKS, QUERY_SEARCH_PARAM } from "src/utils/constants";
